hanging protocol
